set instead of update
update was harder to handle in state, set is idempotent so easier. set gives you the same flexibility but is a bit more unsafe for a user. I think we can make the update cmd line command just check if the hash is there or not, if it is there then set if not it will fail