Added exclusions
Why? Earlier we believed we would not need it because we could just create a new variant which excludes something. However, this make things super messy because we would have to go around in other repositories creating new variants. If you update, you would have to create new variant each time. All in all it would create lots of extra variants and therefore we (well, it is only me now, but...) consider it to be a good feature. Also it helps us with importing data (Ivy I am looking at you)